Automatically backed up by Learn

+++ lib/create.sql
@@ -8,5 +8,5 @@ CREATE TABLE bears (
     gender TEXT, 
     color TEXT, 
     temperament TEXT, 
-    alive NUMERIC, 
+    alive INTEGER, 
 );
